Cleaned up Javascript and added expand logic to task
